(fix)diamond_weight: Changed decimal precision to 2

+++ database/migrations/2024_04_05_044927_create_inventories_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->double('net_weight')->nullable();
             $table->string('karat')->nullable();
             $table->string('diamond_quality')->nullable();
-            $table->double('diamond_weight')->nullable();
+            $table->decimal('diamond_weight')->nullable();
             $table->double('diamond_pcs')->nullable();
             $table->string('color_stone')->nullable();
             $table->boolean('in_stock')->default(true);
